regenrate key on move undo
